build: stop changing behaviour of standard targets
Don't change standard behaviour of 'distclean' and 'maintainer-clean'.
This is annoying for users which expect the default behaviour of a
standard autoconf/automake project.
See [1] which e.g. is explicit about 'maintainer-clean' *not* removing
anything which is necessary to run configure.

The current behaviour breaks the Debian tools when trying to build twice
in a row, as the 'distclean' step removes too many files.

[1] https://www.gnu.org/prep/standards/html_node/Standard-Targets.html