add option for case-insensitive comparisons
This permits to search for json keys in a case-sensitive way.
The default is "off", as this is against the JSON spec. However,
rsyslog needs this capability to increase usability inside the
variable system.

We add a new API call to switch between case-sensitive and
case-insensitive comparison, with case-sensitive being the default.

closes https://github.com/rsyslog/libfastjson/issues/142