"fix" "jump-misses-init" gcc-8 warning
Actually, that warning is overly conservative, and so we
have changed the code in a somewhat suboptimal way to "fix"
it. In this spots, it's not that bad and we wanted to avoid
totally disabling this warning.

If it were more costly in terms of cleanness, we would probaly
have done that. Just mention it to tell anyone else the
cure is not really a good one, just selected due to the
circumstances in this special case.